refactor(openehr-adl): decompose the printer under print/, narrow the public seam, finalize the module map
printer.rs splits into print/{mod,header,terminology,definition,rules,odin}, one
module per artefact section, with the section modules private and print::print
the whole serializer seam; printed output is unchanged (the round-trip corpus,
opt_spec and the 1.4-conversion print comparisons are the net).

The corpus-oracle accessor regression_tag (+ description_of) moves from assemble
to meta, beside the other read-only artefact-summary accessors.

Seam narrowing: every pub item with no consumer in the tests, app/* or
openehr-its drops to pub(crate)/pub(super) — the aom access/interval helpers,
hrid_to_string, assemble/assemble_adl14, most of validate::conformance, and the
validate::rm type-matching helpers (validate_adl14_rm to pub(super)).
